add plot connectome
